Fix /v1/users/login, add service/tokenService
